Fixes
- Fixed int parsing
- fixed police create screen
- Fixed payment name validation

+++ HRP/Controllers/CivilianController.cs
@@ -34,7 +34,7 @@ namespace HRP.Controllers
             }
             else
             {
-                sin = Int16.Parse(session);
+                sin = Int32.Parse(session);
             }
         }
